refactor: 패키지 구조 리팩터링 (250905)
