+++ src/main/java/com/equip/equiprental/auth/dto/LoginRequestDto.java
@@ -1,0 +1,12 @@
+package com.equip.equiprental.auth.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@Builder
+public class LoginRequestDto {
+    private String username;
+    private String password;
+}
