+++ src/main/java/com/equip/equiprental/auth/dto/LoginResponseDto.java
@@ -1,0 +1,17 @@
+package com.equip.equiprental.auth.dto;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@Builder
+public class LoginResponseDto {
+    private Long memberId;
+    private String username;
+    private String name;
+    private String department;
+    private String email;
+    private String role;
+    private String status;
+}
