--- src/main/java/com/equip/equiprental/common/auth/PrincipalDetails.java
@@ -1,39 +1,0 @@
-package com.equip.equiprental.common.auth;
-
-import com.equip.equiprental.domain.member.Member;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-@Getter
-@RequiredArgsConstructor
-public class PrincipalDetails implements UserDetails {
-
-    private final Member member;
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        Collection<GrantedAuthority> authorities = new ArrayList<>();
-        authorities.add(new GrantedAuthority() {
-            @Override
-            public String getAuthority() {
-                return "ROLE_" + member.getRole().name();
-            }
-        });
-        return authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return member.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return member.getUsername();
-    }
-}
