--- src/main/java/com/equip/equiprental/common/auth/PrincipalDetailsService.java
@@ -1,24 +1,0 @@
-package com.equip.equiprental.common.auth;
-
-import com.equip.equiprental.common.exception.CustomException;
-import com.equip.equiprental.common.exception.ErrorType;
-import com.equip.equiprental.domain.member.Member;
-import com.equip.equiprental.domain.member.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-
-@RequiredArgsConstructor
-public class PrincipalDetailsService implements UserDetailsService {
-
-    private final MemberRepository memberRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Member member = memberRepository.findByUsername(username)
-                .orElseThrow(() -> new CustomException(ErrorType.USER_NOT_FOUND));
-
-        return new PrincipalDetails(member);
-    }
-}
