+++ src/main/java/com/equip/equiprental/common/config/WebConfig.java
@@ -1,0 +1,18 @@
+package com.equip.equiprental.common.config;
+
+import com.equip.equiprental.common.interceptor.RequestTraceIdInterceptor;
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebConfig implements WebMvcConfigurer {
+    private final RequestTraceIdInterceptor requestTraceIdInterceptor;
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(requestTraceIdInterceptor);
+    }
+}
