--- src/main/java/com/equip/equiprental/common/controller/ResponseController.java
@@ -1,18 +1,0 @@
-package com.equip.equiprental.common.controller;
-
-import com.equip.equiprental.common.dto.ResponseDto;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-
-public interface ResponseController {
-    default <T> ResponseEntity<ResponseDto<T>> makeResponseEntity(String traceId, HttpStatus httpStatus, String errorCode, String message, T data) {
-        ResponseDto<T> responseDto = ResponseDto.<T>builder()
-                .id(traceId)
-                .code(httpStatus.value())
-                .errorCode(errorCode)
-                .message(message)
-                .data(data)
-                .build();
-        return ResponseEntity.status(httpStatus).body(responseDto);
-    }
-}
