--- src/main/java/com/equip/equiprental/common/dto/ResponseDto.java
@@ -1,17 +1,0 @@
-package com.equip.equiprental.common.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class ResponseDto<T> {
-    private String id;
-    private int code;
-    private String errorCode;
-    private String message;
-    private T data;
-}
-
