+++ src/main/java/com/equip/equiprental/common/response/ResponseDto.java
@@ -1,0 +1,17 @@
+package com.equip.equiprental.common.response;
+
+import lombok.*;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class ResponseDto<T> {
+    private String id;
+    private int code;
+    private String errorCode;
+    private String message;
+    private T data;
+}
+
