--- src/main/java/com/equip/equiprental/config/WebConfig.java
@@ -1,18 +1,0 @@
-package com.equip.equiprental.config;
-
-import com.equip.equiprental.common.interceptor.RequestTraceIdInterceptor;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@RequiredArgsConstructor
-public class WebConfig implements WebMvcConfigurer {
-    private final RequestTraceIdInterceptor requestTraceIdInterceptor;
-
-    @Override
-    public void addInterceptors(InterceptorRegistry registry) {
-        registry.addInterceptor(requestTraceIdInterceptor);
-    }
-}
