--- src/main/java/com/equip/equiprental/domain/member/MemberRepository.java
@@ -1,12 +1,0 @@
-package com.equip.equiprental.domain.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    boolean existsByUsername(String username);
-    boolean existsByEmail(String email);
-
-    Optional<Member> findByUsername(String username);
-}
