--- src/main/java/com/equip/equiprental/domain/member/MemberRole.java
@@ -1,10 +1,0 @@
-package com.equip.equiprental.domain.member;
-
-import lombok.Getter;
-
-@Getter
-public enum MemberRole {
-    USER,
-    MANAGER,
-    ADMIN;
-}
