--- src/main/java/com/equip/equiprental/domain/member/MemberStatus.java
@@ -1,10 +1,0 @@
-package com.equip.equiprental.domain.member;
-
-import lombok.Getter;
-
-@Getter
-public enum MemberStatus {
-    PENDING,
-    ACTIVE,
-    DELETED
-}
