--- src/main/java/com/equip/equiprental/dto/auth/LoginRequestDto.java
@@ -1,12 +1,0 @@
-package com.equip.equiprental.dto.auth;
-
-import lombok.*;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@Builder
-public class LoginRequestDto {
-    private String username;
-    private String password;
-}
