--- src/main/java/com/equip/equiprental/dto/member/SignUpResponseDto.java
@@ -1,21 +1,0 @@
-package com.equip.equiprental.dto.member;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@AllArgsConstructor
-@Builder
-public class SignUpResponseDto {
-    private Long memberId;
-    private String username;
-    private String name;
-    private String department;
-    private String email;
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd'T'HH:mm:ss'Z'", timezone = "UTC")
-    private LocalDateTime createdAt;
-}
