+++ src/main/java/com/equip/equiprental/member/domain/Member.java
@@ -1,0 +1,52 @@
+package com.equip.equiprental.member.domain;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@Table(name="member_tbl")
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class Member {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name="member_id")
+    private Long memberId;
+
+    private String username;
+    private String password;
+    private String name;
+    private String department;
+    private String email;
+
+    @Enumerated(EnumType.STRING)
+    private MemberRole role;
+
+    @Enumerated(EnumType.STRING)
+    private MemberStatus status;
+
+    @Column(name="created_at")
+    private LocalDateTime createdAt;
+
+    @PrePersist
+    public void prePersist() {
+        if (this.createdAt == null) {  // null 체크 꼭 하기
+            this.createdAt = LocalDateTime.now();
+        }
+    }
+
+    @Column(name="updated_at")
+    private LocalDateTime updatedAt;
+
+    @PreUpdate
+    public void preUpdate() {
+        this.updatedAt = LocalDateTime.now();
+    }
+}
