+++ src/main/java/com/equip/equiprental/member/domain/MemberRole.java
@@ -1,0 +1,10 @@
+package com.equip.equiprental.member.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum MemberRole {
+    USER,
+    MANAGER,
+    ADMIN;
+}
