+++ src/main/java/com/equip/equiprental/member/dto/SignUpRequestDto.java
@@ -1,0 +1,19 @@
+package com.equip.equiprental.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@Builder
+public class SignUpRequestDto {
+    private String username;
+    private String password;
+    private String confirmPassword;
+    private String name;
+    private String department;
+    private String email;
+}
