+++ src/main/java/com/equip/equiprental/member/repository/MemberRepository.java
@@ -1,0 +1,13 @@
+package com.equip.equiprental.member.repository;
+
+import com.equip.equiprental.member.domain.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    boolean existsByUsername(String username);
+    boolean existsByEmail(String email);
+
+    Optional<Member> findByUsername(String username);
+}
